deep-research-assistant/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-assistant"
version = "0.1.0"
description = "Multi-agent orchestration system for intelligent research with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Deep Research Assistant Team"}
]
keywords = ["research", "ai", "multi-agent", "langgraph", "rag", "web-scraping"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Researchers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi",
    "uvicorn",
    "langgraph",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "pgvector",
    "psycopg2-binary",
    "openai",
    "tiktoken",
    "aiohttp",
    "playwright",
    "beautifulsoup4",
    "readability-lxml",
    "urllib-robotparser",
    "langdetect",
    "sentence-transformers",
    "pydantic",
    "numpy",
    "pandas",
    "python-dotenv",
    "structlog",
    "asyncio-throttle",
    "aiofiles",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "black",
    "isort",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/deep-research-assistant/deep-research-assistant"
Repository = "https://github.com/deep-research-assistant/deep-research-assistant"
Issues = "https://github.com/deep-research-assistant/deep-research-assistant/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true